# docs/openapi.yaml
openapi: 3.1.0
info:
  title: SubDub API
  version: 1.0.0
  description: |
    OpenAPI contract for the SubDub service.

    Notes:
    - Base API prefix is /api/v1.
    - Arcjet security middleware is applied globally and can return 403/429 before route handlers.
    - Some endpoints are placeholders or have known behavior gaps; see endpoint descriptions.
servers:
  - url: http://localhost:5500
    description: Local development
  - url: '{SERVER_URL}'
    description: Configured deployment URL
    variables:
      SERVER_URL:
        default: http://localhost:5500
...
paths:
...
  /api/v1/subscriptions:
    get:
      tags: [Subscriptions]
      summary: List subscriptions
      operationId: getSubscriptions
      parameters:
        - $ref: '#/components/parameters/PageQuery'
        - $ref: '#/components/parameters/LimitQuery'
      responses:
        '200':
          description: Subscriptions fetched
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionListSuccessResponse'
        '403':
          $ref: '#/components/responses/ArcjetForbidden'
        '429':
          $ref: '#/components/responses/ArcjetRateLimited'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags: [Subscriptions]
      summary: Create subscription
      operationId: createSubscription
      description: Creates a subscription for the authenticated user and triggers reminder workflow.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubscriptionCreateRequest'
      responses:
        '201':
          description: Subscription created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionCreateSuccessResponse'
        '400':
          $ref: '#/components/responses/ValidationOrBusinessError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/ArcjetForbidden'
        '429':
          $ref: '#/components/responses/ArcjetRateLimited'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/v1/subscriptions/user/{id}:
    parameters:
      - $ref: '#/components/parameters/ObjectId'
    get:
      tags: [Subscriptions]
      summary: List subscriptions by user ID
      operationId: getUserSubscriptions
      description: Allowed only if authenticated user ID matches path ID.
      parameters:
        - $ref: '#/components/parameters/PageQuery'
        - $ref: '#/components/parameters/LimitQuery'
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User subscriptions fetched
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionListSuccessResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/ArcjetForbidden'
        '429':
          $ref: '#/components/responses/ArcjetRateLimited'
  /api/v1/subscriptions/{id}/cancel:
    parameters:
      - $ref: '#/components/parameters/ObjectId'
    put:
      tags: [Subscriptions]
      summary: Cancel subscription
      operationId: cancelSubscription
      description: Allowed only if authenticated user owns the subscription.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Subscription cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionSuccessResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/subscriptions/upcoming-renewals:
    get:
      tags: [Subscriptions]
      summary: Upcoming renewals (known implementation issue)
      operationId: getUpcomingRenewals
      parameters:
        - $ref: '#/components/parameters/PageQuery'
        - $ref: '#/components/parameters/LimitQuery'
      description: |
        Intended behavior: returns active subscriptions approaching renewal for authenticated user.

        Current code has route-order and parameter mismatch issues, so runtime behavior does not match intent.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Upcoming renewals fetched (intended)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionListSuccessResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ObjectId:
      name: id
      in: path
      required: true
      description: MongoDB ObjectId
      schema:
        type: string
        pattern: '^[a-fA-F0-9]{24}$'
    PageQuery:
      name: page
      in: query
      required: false
      description: Page number for pagination.
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitQuery:
      name: limit
      in: query
      required: false
      description: Page size for pagination. Values above 100 are capped to 100.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
  responses:
    Unauthorized:
      description: Authentication or authorization failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AuthErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ValidationOrBusinessError:
      description: Validation or business rule error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalServerError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ArcjetForbidden:
      description: Request denied by Arcjet security middleware
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ArcjetErrorResponse'
    ArcjetRateLimited:
      description: Request rate-limited by Arcjet
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ArcjetErrorResponse'
  schemas:
...
    Subscription:
      type: object
      required:
        - name
        - price
        - currency
        - frequency
        - category
        - startDate
        - paymentMethod
        - user
      properties:
        _id:
          type: string
          pattern: '^[a-fA-F0-9]{24}$'
        name:
          type: string
          minLength: 2
          maxLength: 100
        price:
          type: number
          minimum: 0
        currency:
          type: string
          enum: [USD, EUR, GBP, INR, JPY, AUD]
        frequency:
          type: string
          enum: [monthly, quarterly, yearly]
        category:
          type: string
          enum: [Entertainment, Food, Health, Other]
        startDate:
          type: string
          format: date-time
        renewalDate:
          type: string
          format: date-time
        paymentMethod:
          type: string
        status:
          type: string
          enum: [active, cancelled, expired]
        user:
          type: string
          pattern: '^[a-fA-F0-9]{24}$'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        __v:
          type: integer
      additionalProperties: true
    SubscriptionCreateRequest:
      type: object
      required:
        - name
        - price
        - currency
        - frequency
        - category
        - startDate
        - paymentMethod
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 100
          example: Netflix Premium
        price:
          type: number
          minimum: 0
          example: 15.99
        currency:
          type: string
          enum: [USD, EUR, GBP, INR, JPY, AUD]
          example: USD
        frequency:
          type: string
          enum: [monthly, quarterly, yearly]
          example: monthly
        category:
          type: string
          enum: [Entertainment, Food, Health, Other]
          example: Entertainment
        startDate:
          type: string
          format: date-time
          example: '2026-03-01T00:00:00.000Z'
        renewalDate:
          type: string
          format: date-time
          description: Optional; auto-computed if missing.
          example: '2026-03-31T00:00:00.000Z'
        paymentMethod:
          type: string
          example: Visa **** 4242
...
    SubscriptionSuccessResponse:
      type: object
      required: [success, message, data]
      properties:
        success:
          type: boolean
          const: true
        message:
          type: string
          example: Subscription fetched successfully
        data:
          $ref: '#/components/schemas/Subscription'
    SubscriptionListSuccessResponse:
      type: object
      required: [success, message, data, pagination]
      properties:
        success:
          type: boolean
          const: true
        message:
          type: string
          example: Subscriptions fetched successfully
        data:
          type: array
          items:
            $ref: '#/components/schemas/Subscription'
        pagination:
          $ref: '#/components/schemas/PaginationMeta'
    PaginationMeta:
      type: object
      required: [total, page, limit, totalPages, hasNextPage, hasPrevPage]
      properties:
        total:
          type: integer
          minimum: 0
          example: 57
        page:
          type: integer
          minimum: 1
          example: 2
        limit:
          type: integer
          minimum: 1
          maximum: 100
          example: 10
        totalPages:
          type: integer
          minimum: 1
          example: 6
        hasNextPage:
          type: boolean
          example: true
        hasPrevPage:
          type: boolean
          example: true
    SubscriptionCreateSuccessResponse:
      type: object
      required: [success, message, data]
      properties:
        success:
          type: boolean
          const: true
        message:
          type: string
          example: Subscription created successfully
        data:
          type: object
          required: [subscription, workflowRunId]
          properties:
            subscription:
              $ref: '#/components/schemas/Subscription'
            workflowRunId:
              type: string
    ErrorResponse:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
          const: false
        message:
          type: string
        error:
          type: string
          description: Stringified error object in current middleware implementation.
    AuthErrorResponse:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
          const: false
        message:
          type: string
          example: Not authorized, no token
        error:
          type: string
          example: jwt malformed
    ArcjetErrorResponse:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          const: false
        error:
          type: string
          examples: [Too Many Requests, 'Forbidden: Bot detected', Forbidden]